Fix another ESESC_TRACE_LOAD problem
Register $0 can be used as a destination register for loads.
This register does not have a TCG variable associated with it
so it must be handled explicitly.

It's not clear whether the "data" value being passed to ESESC_TRACE_LOAD
is intended to be the final register value or the data read from memory.